requirements.txt
asgiref==3.8.1
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.15.4
mysqlclient==2.2.4
platformdirs==4.2.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.3
whitenoise==6.6.0
